moved godcat files to its own folder
