+++ NPCs/Bosses/Godcat/Godcat_Dark.png
