+++ NPCs/Bosses/Godcat/Godcat_DarkBall_Small.png
