+++ NPCs/Bosses/Godcat/Godcat_DarkBlade.png
