+++ NPCs/Bosses/Godcat/Godcat_Destroyer_Attack.png
