+++ NPCs/Bosses/Godcat/Godcat_Light.png
