+++ NPCs/Bosses/Godcat/Godcat_LightBall.png
