+++ NPCs/Bosses/Godcat/Godcat_LightReturnBall.png
