+++ NPCs/Bosses/Godcat/Godcat_Light_Head_Boss.png
